pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100
skip-string-normalization = true
target-version = ["py38"]
# old versions of black used to put a comma at the end of
#
# fn(
#     arg1, arg2,
# )
#
# and new versions interpret this as a "magic comma" and use it to split it to one item per line
# (https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#the-magic-trailing-comma):
#
# fn(
#     arg1,
#     arg2,
# )
#
# which we do not like. So we preserve the old behavior with the following
#
# We could also keep the new magic comma behavior, but do one-off reformatting using the following
# setting just to remove the magic commas in the first place
skip-magic-trailing-comma = true

[tool.isort]
line_length = 100
profile = "black"

[tool.poetry]
name = "celus"
version = "5.1.0"
description = "web application for harvesting and visualization of usage data of electronic information sources"
authors = ["Beda Kosata <beda@bigdigdata.com>", "Stepan Henek <stepan@bigdigdata.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8.8"

celery = "~5.2.3"
celus-nibbler = { git = "https://github.com/Big-Dig-Data/celus-nibbler.git", tag="v7.4.0" }
celus-nigiri = { git = "https://github.com/Big-Dig-Data/celus-nigiri.git", tag="v1.2.2" }
colorlog = "^6.7.0"
diskcache = "5.4"
dj-rest-auth = "~2.2.5"
Django = "~3.2.7"
django-allauth = "~0.51.0"
django-anymail = "~9.0"
django-cachalot = "~2.5.2"
django-celery-results = "~2.4.0"
django-colorfield = "^0.8.0"
django-impersonate =  "~1.8.2"
django-import-export = "^3.0.2"
django-modeltranslation = "~0.18.7"
django-mptt = "~0.14.0"
django-postgres-copy = "^2.7.0"
django-prometheus = "~2.2.0"
django-redis = "~5.2.0"
django-reversion = "~5.0.4"
django-test-migrations = "^1.1.0"
djangorestframework = "~3.14.0"
djangorestframework-api-key = "^2.0.0"
drf-nested-routers = "~0.93.4"
hcube = "0.7.1"
isbnlib = "^3.10.3"
lz4 = "~4.0.2"
openpyxl = "~3.0.9"
Pillow = "^9.3.0"
psutil = "~5.9.4"
psycopg2-binary = "~2.9.1"
pycounter = {git = "https://github.com/beda42/pycounter.git", rev = "7d3b8470d909e3a9c989492c01fd8d40b89248cd"}
pytest-cov = "^2.11.1"
pytest-xdist = "^2.2.1"
python-decouple = "^3.3"
python-magic = "~0.4.15"
pytz = "*"
redis = "~4.4.0"
requests = "^2.23.0"
rest-pandas = "~1.1.0"
sentry-sdk = "^1.12.1"
toml = "^0.10.2"
urllib3 = "~1.26.13"
XlsxWriter = "^3.0.3"


[tool.poetry.dev-dependencies]
black = "^22.12.0"
isort = "^5.11.3"
dj-inmemorystorage = "^2.1.0"
django-debug-toolbar = "^3.2.1"
django-extensions = "^3.1.3"
django-querycount = "^0.7.0"
django-webtest = "^1.9.7"
factory-boy = "^3.2.1"
Faker = "^15.3.4"
freezegun = "^1.2.2"
pre-commit = "^2.4.0"
pydot = "^1.4.2"
pyfakefs = "^5.0.0"
pyparsing = "^2.4.7"
pytest = "^7.2.0"
pytest-django = "^4.2.0"
pytest-env = "^0.8.1"
requests-mock = "~1.10.0"
setuptools = "^65.6.3"
sphinx = "^5.3.0"
watchdog = {extras = ["watchmedo"], version = "^2.1.6"}
flake8 = "^6.0.0"
